Simple Interpolation Ratio + Area to Triangles
Added the basics of the interpolation ratio: Now we can get the Barcyntric
interpolation from points inside the screen_image. Next will be
implementing the Z-Buffering algorithm and then the Vertices' Attributes
interpolation.

Now you can get the traingle area, which is formed when the triangle is
created by using the determinants of the vectors.